+++ src/index.ts
@@ -1,5 +1,4 @@
-export type Mode = "instance" | "deploy" | "fixtures" | "validate";
+export type Mode = "test" | "deploy" | "fixtures" | "validate";
 
 export * from "./context";
-export * from "./instance";
 export * from "./kivik";